Updated the units manager to use the new data access system
